// front/gapi_calendar.d.ts
  export class calendarList {
    static list(parameters?: CalendarListListParameters): HttpRequest<CalendarList>;
    static insert(parameters: CalendarListInsertParameters): HttpRequest<CalendarListEntry>;
  }
// ...
  interface CalendarListListParameters {
    maxResults?: integer | undefined;
    // The minimum access role for the user in the returned entries. Optional. The default is no restriction. Acceptable values are:
    minAccessRole?: AccessRoleWithoutNone | undefined;
    pageToken?: string | undefined;
    showDeleted?: boolean | undefined;
    showHidden?: boolean | undefined;
    syncToken?: string | undefined;
  }

  interface CalendarListInsertParameters {
    // Parameters
    // Optional query parameters
    colorRgbFormat?: boolean | undefined;

    // CalendarList resource
    resource: CalendarListInput;
  }
// ...
  interface CalendarList {
    kind: 'calendar#calendarList';
    etag: etag;

    /**
     * Token used to access the next page of this result.
     * Omitted if no further results are available, in which case nextSyncToken is provided.
     */
    nextPageToken?: string | undefined;

    /**
     * Token used at a later point in time to retrieve only the entries that have changed since this result was returned.
     * Omitted if further results are available, in which case nextPageToken is provided.
     */
    nextSyncToken?: string | undefined;

    items: Event[];
  }
// ...
  interface CalendarListEntry {
    kind: 'calendar#calendarListEntry';
    etag: etag;
    id: string;
    summary: string;
    description?: string | undefined;
    location?: string | undefined;
    timeZone?: string | undefined;
    summaryOverride?: string | undefined;
    colorId?: string | undefined;
    backgroundColor?: string | undefined;
    foregroundColor?: string | undefined;
    hidden?: boolean | undefined;
    selected?: boolean | undefined;
    // The effective access role that the authenticated user has on the calendar. Read-only.
    accessRole: AccessRoleWithoutNone;
    defaultReminders: {
      method: ReminderMethod;
      minutes: integer;
    }[];
    notificationSettings?: {
      notifications: {
        type: NotificationType;
        method: string;
      }[];
    } | undefined;
    primary?: boolean | undefined;
    deleted?: boolean | undefined;
  }
// ...
  type etag = string;